files renamed / code cleanup
