+++ frontend/src/components/Avionics/AvionicsHero.js
