+++ frontend/src/components/Avionics/AvionicsIcons.js
