+++ frontend/src/components/Avionics/AvionicsImg.js
