+++ frontend/src/components/Flir/FlirHero.js
