+++ frontend/src/components/Flir/FlirTable.js
