+++ frontend/src/components/Flir/FlirText.js
