+++ frontend/src/components/Garmin.js
@@ -7,8 +7,6 @@ import ProductCarousel from "./ProductCarousel";
 
 
 const Garmin = () => {
-
-
     return (
         <Row className='d-flex justify-content-center Garmin_card-row bg-white'>
             <Col md={6} xs={12} className='d-flex align-content-center m-0 p-0'>
