+++ frontend/src/screens/AdminScreens/UserEditScreen.js
@@ -7,7 +7,7 @@ import Loader from '../../components/Loader'
 import FormContainer from "../../components/FormContainer";
 import { getUserDetails, updateUser } from '../../actions/userActions'
 import { USER_UPDATE_RESET} from "../../constants/userConstants";
-import StickyHeader from "../../components/Headers-Nav-Footer/StickyHeader";
+import StickyHeader from "../../components/Header/StickyHeader";
 
 const UserEditScreen = ({ match, history}) => {
     const userId = match.params.id;
