+++ frontend/src/screens/AdminScreens/UserListScreen.js
@@ -5,7 +5,7 @@ import {useDispatch, useSelector} from 'react-redux'
 import Message from '../../components/Message'
 import Loader from '../../components/Loader'
 import { listUsers, deleteUser } from "../../actions/userActions"
-import StickyHeader from "../../components/Headers-Nav-Footer/StickyHeader";
+import StickyHeader from "../../components/Header/StickyHeader";
 
 const UserListScreen = ({ history }) => {
     const dispatch = useDispatch();
