+++ frontend/src/screens/Payment/PlaceOrderScreen.js
@@ -3,7 +3,7 @@ import {Link} from "react-router-dom";
 import {Button, Row, Col, ListGroup, Image, Card} from "react-bootstrap";
 import {useDispatch, useSelector} from "react-redux";
 import Message from "../../components/Message";
-import CheckoutSteps from "../../components/CheckoutSteps";
+import CheckoutSteps from "../../components/Checkout/CheckoutSteps";
 import {createOrder} from "../../actions/orderActions";
 
 const PlaceOrderScreen = ({ history }) => {
