+++ frontend/src/screens/Payment/ShippingScreen.js
@@ -2,7 +2,7 @@ import React, {useState} from 'react'
 import {Form, Button} from 'react-bootstrap'
 import {useDispatch, useSelector} from 'react-redux'
 import FormContainer from "../../components/FormContainer";
-import CheckoutSteps from "../../components/CheckoutSteps";
+import CheckoutSteps from "../../components/Checkout/CheckoutSteps";
 import { saveShippingAddress } from "../../actions/cartActions";
 
 const ShippingScreen = ({ history }) => {
